move some files out of top level folders
