+++ frontend/src/modules/app/OverlayScreen.tsx
@@ -18,7 +18,7 @@ import {
 import { MessageType } from '../../store/proto/types';
 
 import ModalPopup from '../shared/ModalPopup';
-import MultiStepWizard from '../modals/MultiStepWizard';
+import MultiStepWizard from './multi-step-wizard/MultiStepWizard';
 import { getScreenLockPopup, setScreenLockPopup } from './Service';
 import { updateState } from '../../store/controller/actions';
 import { getBackendConnected } from '../../store/connection/selectors';
