+++ frontend/src/modules/app/toolbar/ToolBar.tsx
@@ -10,7 +10,7 @@ import React from 'react';
 import { useLocation } from 'react-router-dom';
 import ViewDropdown from '../../dashboard/views/ViewDropdown';
 import ClockIcon from '../../icons/ClockIcon';
-import ModesDropdown from '../../modes/ModesDropdown';
+import ModesDropdown from './ModesDropdown';
 import { DASHBOARD_ROUTE, QUICKSTART_ROUTE } from '../navigation/constants';
 import Alarms from './Alarms';
 import StartButton, { StartButtonModalPopup } from './StartPauseButton';
