+++ frontend/src/modules/event-logs/EventLog.tsx
