+++ frontend/src/modules/event-logs/EventType.tsx
