+++ frontend/src/modules/event-logs/LogsPage.tsx
