+++ frontend/src/modules/event-logs/constants.ts
