+++ frontend/src/modules/event-logs/container/EventlogDetails.tsx
