--- frontend/src/modules/modes/index.ts
@@ -1,1 +1,0 @@
-export * from './ModeBanner';
